+++ backend/app/models/position.py
@@ -2,6 +2,7 @@
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     Date,
@@ -41,34 +42,34 @@ class Position(Base):
     position_type: Mapped[str] = mapped_column(String(10), nullable=False)
     asset_symbol: Mapped[str | None] = mapped_column(String(20), nullable=True)
     asset_address: Mapped[str | None] = mapped_column(String(100), nullable=True)
-    current_shares_or_amount: Mapped[float | None] = mapped_column(
+    current_shares_or_amount: Mapped[Decimal | None] = mapped_column(
         Numeric(30, 12), nullable=True
     )
-    cost_basis_fifo_usd: Mapped[float | None] = mapped_column(
+    cost_basis_fifo_usd: Mapped[Decimal | None] = mapped_column(
         Numeric(18, 2), nullable=True
     )
-    cost_basis_wac_usd: Mapped[float | None] = mapped_column(
+    cost_basis_wac_usd: Mapped[Decimal | None] = mapped_column(
         Numeric(18, 2), nullable=True
     )
-    total_yield_earned_usd: Mapped[float] = mapped_column(
+    total_yield_earned_usd: Mapped[Decimal] = mapped_column(
         Numeric(18, 2), default=0, nullable=False
     )
-    total_borrow_cost_usd: Mapped[float] = mapped_column(
+    total_borrow_cost_usd: Mapped[Decimal] = mapped_column(
         Numeric(18, 2), default=0, nullable=False
     )
-    unrealised_pnl_usd: Mapped[float | None] = mapped_column(
+    unrealised_pnl_usd: Mapped[Decimal | None] = mapped_column(
         Numeric(18, 2), nullable=True
     )
-    unrealised_pnl_native: Mapped[float | None] = mapped_column(
+    unrealised_pnl_native: Mapped[Decimal | None] = mapped_column(
         Numeric(24, 8), nullable=True
     )
-    realised_pnl_fifo_usd: Mapped[float] = mapped_column(
+    realised_pnl_fifo_usd: Mapped[Decimal] = mapped_column(
         Numeric(18, 2), default=0, nullable=False
     )
-    realised_pnl_wac_usd: Mapped[float] = mapped_column(
+    realised_pnl_wac_usd: Mapped[Decimal] = mapped_column(
         Numeric(18, 2), default=0, nullable=False
     )
-    health_factor: Mapped[float | None] = mapped_column(Numeric(8, 4), nullable=True)
+    health_factor: Mapped[Decimal | None] = mapped_column(Numeric(8, 4), nullable=True)
     status: Mapped[str] = mapped_column(String(10), default="active", nullable=False)
     closed_at: Mapped[datetime | None] = mapped_column(
         DateTime(timezone=True), nullable=True
@@ -82,16 +83,16 @@ class Position(Base):
     manual_protocol_name: Mapped[str | None] = mapped_column(Text, nullable=True)
     pendle_position_type: Mapped[str | None] = mapped_column(String(5), nullable=True)
     pendle_maturity_date: Mapped[Date | None] = mapped_column(Date, nullable=True)
-    pendle_implied_apy_at_entry: Mapped[float | None] = mapped_column(
+    pendle_implied_apy_at_entry: Mapped[Decimal | None] = mapped_column(
         Numeric(6, 3), nullable=True
     )
-    pendle_yt_claimed_usd: Mapped[float | None] = mapped_column(
+    pendle_yt_claimed_usd: Mapped[Decimal | None] = mapped_column(
         Numeric(18, 2), nullable=True
     )
-    pendle_yt_pending_usd: Mapped[float | None] = mapped_column(
+    pendle_yt_pending_usd: Mapped[Decimal | None] = mapped_column(
         Numeric(18, 2), nullable=True
     )
-    pendle_pt_maturity_value_usd: Mapped[float | None] = mapped_column(
+    pendle_pt_maturity_value_usd: Mapped[Decimal | None] = mapped_column(
         Numeric(18, 2), nullable=True
     )
     created_at: Mapped[datetime] = mapped_column(
