+++ backend/app/models/vault_concentration.py
@@ -1,5 +1,7 @@
 """VaultConcentration model — whale concentration per vault (§10, hypertable)."""
 
+from decimal import Decimal
+
 from sqlalchemy import DateTime, Integer, Numeric, String
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
@@ -18,6 +20,6 @@ class VaultConcentration(Base):
         DateTime(timezone=True), primary_key=True, nullable=False
     )
     top_n: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    top_n_pct: Mapped[float | None] = mapped_column(Numeric(5, 2), nullable=True)
+    top_n_pct: Mapped[Decimal | None] = mapped_column(Numeric(5, 2), nullable=True)
     top_holders: Mapped[dict | None] = mapped_column(JSONB, nullable=True)
     total_holders: Mapped[int | None] = mapped_column(Integer, nullable=True)
